+++ src/chunked_scatter/chunked_scatter.py
@@ -22,7 +22,7 @@ import argparse
 from pathlib import Path
 from typing import Generator, Iterable, List
 
-from .parsers import BedRegion, file_to_regions
+from .parsers import BedRegion, SUPPORTED_EXTENSIONS_STRING, file_to_regions
 
 
 def region_chunker(regions: Iterable[BedRegion], chunk_size: int, overlap: int
@@ -57,7 +57,8 @@ def region_chunker(regions: Iterable[BedRegion], chunk_size: int, overlap: int
 def chunked_scatter(regions: Iterable[BedRegion],
                     chunk_size: int,
                     overlap: int,
-                    minimum_base_pairs: int,
+                    list_size: int,
+                    size_is_maximum: bool = False,
                     contigs_can_be_split: bool = False,
                     ) -> Generator[List[BedRegion], None, None]:
     """
@@ -67,8 +68,9 @@ def chunked_scatter(regions: Iterable[BedRegion],
     :param regions: The regions which to chunk.
     :param chunk_size: The size of each chunk.
     :param overlap: How much overlap there should be between chunks.
-    :param minimum_base_pairs: What the minimum amount of base pairs should be
+    :param list_size: What the minimum amount of base pairs should be
     that the regions encompass per List.
+    :param size_is_maximum: Use list_size as a maximum instead of a minimum
     :param contigs_can_be_split: Whether contigs (chr1, for example) are
     allowed to be split across multiple lists.
     :return: Lists of BedRegions, which can be converted into BED files.
@@ -80,14 +82,17 @@ def chunked_scatter(regions: Iterable[BedRegion],
         # If the next chunk is on a different contig
         if contigs_can_be_split or chunk.contig != current_contig:
             current_contig = chunk.contig
-            # and the current bed file contains enough bases
-            if current_scatter_size >= minimum_base_pairs:
+            # Yield if the minimum is reached, or if current chunk will
+            # overflow the size and there is at least one chunk already.
+            size_to_check = (current_scatter_size + len(chunk)
+                             if size_is_maximum else current_scatter_size)
+            if size_to_check >= list_size and len(chunk_list) > 0:
                 yield chunk_list
                 chunk_list = []
                 current_scatter_size = 0
         # Add the chunk to the bed file
         chunk_list.append(chunk)
-        current_scatter_size += (chunk.end-chunk.start)
+        current_scatter_size += len(chunk)
     # If there are leftovers yield them.
     if chunk_list:
         yield chunk_list
@@ -124,11 +129,10 @@ def common_parser() -> argparse.ArgumentParser:
                         help="The prefix of the ouput files. Output will be "
                         "named like: <PREFIX><N>.bed, in which N is an "
                         "incrementing number. Default 'scatter-'.")
-    parser.add_argument("input", metavar="INPUT", type=Path,
-                        help="The input file, either a bed file or a sequence "
-                        "dict. Which format is used is detected by the "
-                        "extension: '.bed', '.fai' or '.dict'. This option is "
-                        "mandatory.")
+    parser.add_argument("input", metavar="INPUT", type=str,
+                        help=f"The input file. The format is detected by the "
+                             f"extension. Supported extensions are: "
+                             f"{SUPPORTED_EXTENSIONS_STRING}.")
     parser.add_argument("-S", "--split-contigs", action="store_true",
                         help="If set, contigs are allowed to be split up over "
                              "multiple files.")
@@ -177,7 +181,8 @@ def main():
     scattered_chunks = chunked_scatter(file_to_regions(args.input),
                                        args.chunk_size, args.overlap,
                                        args.minimum_bp_per_file,
-                                       args.split_contigs)
+                                       size_is_maximum=False,
+                                       contigs_can_be_split=args.split_contigs)
     out_files = region_lists_to_scatter_files(scattered_chunks, args.prefix)
     if args.print_paths:
         print("\n".join(out_files))